+++ src/non-copyable.h
@@ -28,5 +28,7 @@ protected:
 private:
     NonCopyable(const NonCopyable&) = delete;
     NonCopyable& operator=(const NonCopyable&) = delete;
+    NonCopyable(NonCopyable&&) = delete;
+    NonCopyable& operator=(NonCopyable&&) = delete;
 };
 #endif /* NonCopyable_h */