+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Updates
 - New format specifier `%session_id`
 - Client integrity tasks are paused for specific clients being processed instead of all clients
+- Unknown client trying to use known logger using valid access code is no longer allowed
 
 ## [1.6.0] - 27-02-2018
 ### Updates